requirements.txt
numpy==2.1.2
openai==1.52.0
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
streamlit==1.39.0
